\name{Beta}
\alias{Beta}
\title{Internal function, used to compare covariance matrices.}
\description{Internal function}
\usage{
Beta(nk, etak, Sk)
}
\arguments{
  \item{nk}{scalar, number of samples,}
  \item{etak}{scalar, number of degrees of freedom}
  \item{Sk}{square matrix, homogeneous to a covariance matrix.}
}
\keyword{internal}