\name{gcdalin}
\alias{gcdalin}
\title{
Linear gCDA
}
\description{
Internal function used by the function gcda.
}
\usage{
gcdalin(X, y, gr, alpha = 0.5)
}
\references{
This function is inspired from the function lda (package MASS)
}
\author{
Vincent Guillemot
}
\seealso{
\code{\link{gcda}},\code{\link{gcdaquad}}
}
\keyword{internal}