\name{gcdaquad}
\alias{gcdaquad}
\title{
Quadratic gCDA
}
\description{
Internal function used by the function gcda.
}
\usage{
gcdaquad(X, y, gr1, gr2, alpha1 = 0.5, alpha2 = 0.5)
}
\references{
This function is inspired from the function lda (package MASS)
}
\author{
Vincent Guillemot
}
\seealso{
\code{\link{gcda}},\code{\link{gcdalin}}
}
\keyword{internal}