\name{predictgcdalin}
\alias{predictgcdalin}
\title{
Linear gCDA, prediction function
}
\description{
Internal function used by the function gcda.
}
\usage{
predictgcdalin(X, y, Xnew, ynew, gr, alpha)
}
\references{
This function is inspired from the function lda (package MASS)
}
\author{
Vincent Guillemot
}

\seealso{
\code{\link{gcda}},\code{\link{predictgcdaquad}}
}
\keyword{internal}