\name{predictgcdaquad}
\alias{predictgcdaquad}
\title{
Quadratic gCDA, prediction function
}
\description{
Internal function used by the function gcda.
}
\usage{
predictgcdaquad(X, y, Xnew, ynew, gr1, gr2, alpha1, alpha2)
}
\references{
This function is inspired from the function lda (package MASS)
}
\author{
Vincent Guillemot
}
\seealso{
\code{\link{gcda}},\code{\link{predictgcdalin}}
}
\keyword{internal}